+++ x/distribution/types/events.go
@@ -12,4 +12,5 @@ const (
 	AttributeKeyWithdrawAddress = "withdraw_address"
 	AttributeKeyValidator       = "validator"
 	AttributeKeyDelegator       = "delegator"
+	AttributeValueCategory      = ModuleName
 )